# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CareerProfileInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  degree: string;
  /** @minLength 2 */
  school: string;
  /** @minItems 1 */
  skills: string[];
  interests: string[];
  /** @nullable */
  resume: string | null;
}

export interface CareerProgressInput {
  completedPrep: string[];
  savedProjects: string[];
}

export interface CareerMatch {
  role: string;
  company: string;
  score: number;
  reason: string;
}

export interface SkillGap {
  name: string;
  current: number;
  required: number;
  label: string;
}

export interface RoadmapStep {
  week: string;
  title: string;
  detail: string;
}

export interface CareerProject {
  id: string;
  name: string;
  detail: string;
  skills: string;
  time: string;
  level: string;
}

export interface InterviewItem {
  id: string;
  title: string;
  detail: string;
}

export interface JobRole {
  role: string;
  detail: string;
}

export interface CareerDashboard {
  id: string;
  profile: CareerProfileInput;
  matches: CareerMatch[];
  skillGaps: SkillGap[];
  roadmap: RoadmapStep[];
  projects: CareerProject[];
  interviewChecklist: InterviewItem[];
  jobRoles: JobRole[];
  completedPrep: string[];
  savedProjects: string[];
  createdAt: string;
  updatedAt: string;
}