# lib/api-zod/src/generated/types/careerDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareerMatch } from './careerMatch';
import type { CareerProfileInput } from './careerProfileInput';
import type { CareerProject } from './careerProject';
import type { InterviewItem } from './interviewItem';
import type { JobRole } from './jobRole';
import type { RoadmapStep } from './roadmapStep';
import type { SkillGap } from './skillGap';

export interface CareerDashboard {
  id: string;
  profile: CareerProfileInput;
  matches: CareerMatch[];
  skillGaps: SkillGap[];
  roadmap: RoadmapStep[];
  projects: CareerProject[];
  interviewChecklist: InterviewItem[];
  jobRoles: JobRole[];
  completedPrep: string[];
  savedProjects: string[];
  createdAt: Date;
  updatedAt: Date;
}