# lib/api-zod/src/generated/types/careerMatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerMatch {
  role: string;
  company: string;
  score: number;
  reason: string;
}