# lib/api-zod/src/generated/types/careerProfileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerProfileInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  degree: string;
  /** @minLength 2 */
  school: string;
  /** @minItems 1 */
  skills: string[];
  interests: string[];
  /** @nullable */
  resume: string | null;
}