# lib/api-zod/src/generated/types/careerProgressInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerProgressInput {
  completedPrep: string[];
  savedProjects: string[];
}