# lib/api-zod/src/generated/types/careerProject.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerProject {
  id: string;
  name: string;
  detail: string;
  skills: string;
  time: string;
  level: string;
}