# lib/api-zod/src/generated/types/interviewItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewItem {
  id: string;
  title: string;
  detail: string;
}