# lib/api-zod/src/generated/types/roadmapStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoadmapStep {
  week: string;
  title: string;
  detail: string;
}