# lib/api-zod/src/generated/types/skillGap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillGap {
  name: string;
  current: number;
  required: number;
  label: string;
}